Cleanup and move entries to wasm_entries
