+++ web/src/platform/multithreaded/mod.rs
@@ -1,2 +1,3 @@
 pub mod pool;
 pub mod pool_scheduler;
+pub mod wasm_entries;
