+++ web/src/platform/multithreaded/wasm_entries.rs
@@ -1,0 +1,12 @@
+use wasm_bindgen::prelude::*;
+use wasm_bindgen_futures::JsFuture;
+
+use crate::platform::multithreaded::pool::Work;
+
+/// Entry point invoked by the worker.
+#[wasm_bindgen]
+pub async fn multithreaded_worker_entry(ptr: u32) -> Result<(), JsValue> {
+    let work = unsafe { Box::from_raw(ptr as *mut Work) };
+    JsFuture::from(work.execute()).await?;
+    Ok(())
+}
